+++ Adj/Program/Primitive/Generation.hs
@@ -4,7 +4,7 @@
 module Adj.Program.Primitive.Generation where
 
 import Adj.Auxiliary (Casted, Casting ((=-), (-=)), type (=!?=), FG (FG), FFGH (FFGH), type (=!!??=), Structural (Structural))
-import Adj.Algebra.Category (Semigroupoid ((.)), Category ((.:), (...:), (....:), identity), Functor (map), Functoriality (Natural), Covariant, Bindable, Traversable, Semimonoidal, Component (component), Identity (Identity), Day (Day), type (-->), type (-/->), Flat (Flat), Dual, Kleisli (Kleisli), (->-), (->>-), (->>--), (->--), (-->--), (-/>/-), (-/>>/-), (=-=))
+import Adj.Algebra.Category (Semigroupoid ((.)), Category ((.:), (...:), (....:), identity), Functor (map), Covariant, Bindable, Traversable, Semimonoidal, Component (component), Identity (Identity), Day (Day), type (-->), type (-/->), Flat (Flat), Dual, Kleisli (Kleisli), (->-), (->>-), (->>--), (->--), (-->--), (-/>/-), (-/>>/-), (=-=))
 import Adj.Algebra.Set (Setoid, (:*:) ((:*:)), (:+:) (This, That))
 
 newtype Generation f g o = Generation
@@ -22,14 +22,14 @@ instance Casting (->) (Generation f g) where
 instance
 	( forall o . Functor (-->) (-->) (Flat f o)
 	, forall o . Functor (-->) (-->) (Dual f o)
-	, Covariant Natural Functor (->) (->) g
+	, Covariant Flat Functor g (->) (->)
 	) => Functor (-->) (-->) (Generation f g) where
 	map (Flat m) = Flat . (=-=) . (=-=) .: (->>--) m . (-->--) ((=-=) (m ->>-))
 
 instance
 	( forall o . Functor (-->) (-->) (Flat f o)
 	, forall o . Functor (-->) (-->) (Dual f o)
-	, Covariant Natural Functor (->) (->) g
+	, Covariant Flat Functor g (->) (->)
 	, Semimonoidal Functor f f (-->) (-->) h
 	-- TODO: how did we get to this?
 	, Bindable Functor (->) (->) h
@@ -55,6 +55,6 @@ pattern Load :: o -> Instruction f o
 pattern Load x <- Generation (FFGH (This (Identity x)))
 	where Load x = Generation . FFGH . This .: Identity x
 
-instance Covariant Natural Functor (->) (->) f => Component (-->) f (Instruction f) where
+instance Covariant Flat Functor f (->) (->) => Component (-->) f (Instruction f) where
 	component = Flat .: \x -> Generation . FFGH . That . FG
 		....: Generation . FFGH . This . Identity ->- x